:package: modal directory rearranged
